requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
playwright==1.40.0
python-telegram-bot==20.7
pyyaml==6.0.1
flask==2.3.3
schedule==1.2.1
SQLAlchemy==2.0.23
gunicorn==21.2.0
whitenoise==6.5.0
